# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telco Churn DataOps Pipeline API",
    "description": "Cloud-native, versioned REST API exposing application, deployment, pipeline and dataset details for the AIMLCZG549 Assignment-1 Telco Churn DataOps pipeline.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/pipeline/runs": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Pipeline Runs",
        "operationId": "pipeline_runs_api_v1_pipeline_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipeline/latest": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Pipeline Latest",
        "operationId": "pipeline_latest_api_v1_pipeline_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dataset/info": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Dataset Info",
        "operationId": "dataset_info_api_v1_dataset_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eda/feature-importance": {
      "get": {
        "tags": [
          "eda"
        ],
        "summary": "Feature Importance Endpoint",
        "operationId": "feature_importance_endpoint_api_v1_eda_feature_importance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureImportanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pipeline/runs": {
      "get": {
        "tags": [
          "legacy"
        ],
        "summary": "Pipeline Runs",
        "operationId": "pipeline_runs_api_pipeline_runs_get",
        "deprecated": true,
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetInfoResponse": {
        "properties": {
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "columns": {
            "type": "integer",
            "title": "Columns"
          },
          "column_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Column Names"
          },
          "target_variable": {
            "type": "string",
            "title": "Target Variable"
          },
          "churn_rate": {
            "type": "number",
            "title": "Churn Rate"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "columns",
          "column_names",
          "target_variable",
          "churn_rate"
        ],
        "title": "DatasetInfoResponse"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "NOT_FOUND"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "FeatureImportanceResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "top_feature_importances": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Top Feature Importances"
          },
          "top_correlations_with_churn": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Top Correlations With Churn"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "top_feature_importances",
          "top_correlations_with_churn"
        ],
        "title": "FeatureImportanceResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "timestamp_utc": {
            "type": "string",
            "title": "Timestamp Utc"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_sec": {
            "type": "number",
            "title": "Duration Sec"
          },
          "rows_processed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Processed"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "timestamp_utc",
          "status",
          "duration_sec"
        ],
        "title": "PipelineRun"
      },
      "PipelineRunsResponse": {
        "properties": {
          "total_runs": {
            "type": "integer",
            "title": "Total Runs"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/PipelineRun"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "total_runs",
          "runs"
        ],
        "title": "PipelineRunsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}